+++ client/src/components/Nav/NavBar.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Link, useLocation } from 'react-router-dom'
 import ThemeToggle from '../ThemeToggle/ThemeToggle'
 import './NavBar.css'